Finished writing fetchtypes file, fixed tests and moved into separate tests folder, removed paperscraper folder (not part of software and it is still on the paperscraper branch), organized GitHub directory into logical tree of files (renamed some things), moved data exploration into csvs in the data folder, cleaned up the technical readme in the auto_db_pipeline folder, added the regex scripts for all the protein IDs (cleared up notes/other-datatypes to reflect this, solved the doi redirect to full text for the vaccines and viruses journals
