+++ .gitignore
@@ -1,6 +1,7 @@
 SAbDab-master/
 old_keyword_files/
 .DS_Store
+__pycache__/
 keywords/__pycache__
 keywords/.pytest_cache
 keywords/tests/__pycache__
@@ -9,3 +10,5 @@ paperscraper/.ipynb_checkpoints/
 paperscraper/check_coverage.ipynb
 paperscraper/testfile.ipynb
 data/biorxiv_all*.jsonl
+temp.ipynb
+
