+++ auto_db_pipeline/README.md
@@ -1,21 +1,53 @@
 ## Input / Output
 
-Text here.
+We start with keywords, and use [keywords2papers.py](keywords2papers.py) to get to papers, and then use [papers2info.py](papers2info.py) to get to protein informatics.
 
 ---
 ## Protein ID formatting
-### Genbank protein ids.
+### GenBank protein IDs
 "Three letters followed by five digits, a period, and a version number."
 The rules are listed at:
-* [NCBI SequenceIDs](https://www.ncbi.nlm.nih.gov/genbank/sequenceids/)
-* [NCBI Sitemap](https://www.ncbi.nlm.nih.gov/Sitemap/samplerecord.html)
+* [Rule quoted from genbank/sequenceids](https://www.ncbi.nlm.nih.gov/genbank/sequenceids/)
+* [Also found on sitemap/samplerecord](https://www.ncbi.nlm.nih.gov/Sitemap/samplerecord.html)
 
-#### PDB ids
+### GenBank Accessions
+
+* [All rules found on genbank/acc_prefix](https://www.ncbi.nlm.nih.gov/genbank/acc_prefix/)
+* [Restated on WUSTL.edu](https://community.gep.wustl.edu/repository/course_materials_WU/annotation/Genbank_Accessions.pdf)
+
+#### Nucleotide accessions
+1 letter + 5 digits, or, 2 letters + 6 digits, or, 2 letters + 8 digits.
+#### Protein accessions
+3 letters + 5 digits, or, 3 letters + 7 digits.
+Notice that this first rule is identical to the GenBank protein ID rules, but without the dot version number at the end.
+### RefSeq IDs
+
+"Two letters followed by an underscore bar and six or more digits."
+* [Rule quoted from genbank/samplerecord](https://www.ncbi.nlm.nih.gov/genbank/samplerecord/#:~:text=Records%20from%20the%20RefSeq)
+* [Briefly described by UNMC.edu](https://www.unmc.edu/bsbc/docs/formats.htm)
+* [Example SARS entry](https://www.ncbi.nlm.nih.gov/protein/1796318597)
+
+The two letters indicate which kind of RefSeq the ID is, these are [described by the NCBI here](https://www.ncbi.nlm.nih.gov/books/NBK21091/table/ch18.T.refseq_accession_numbers_and_mole/?report=objectonly).
+
+These are a type of accession.
+
+### GenInfo (GI) Identifiers (GenBank)
+
+* [Rule found on genbank/sequenceids](https://www.ncbi.nlm.nih.gov/genbank/sequenceids)
+* [Also described on sitemap/sequenceIDs](https://www.ncbi.nlm.nih.gov/Sitemap/sequenceIDs.html)
+
+It is a simple series of digits that are assigned consecutively to each sequence record processed by NCBI.
+
+
+### PDB IDs
 1. All characters are alphabetical or numeric
 1. Length of four.
 2. First charater is numeric.
 3. First character is in the range of 1-9 inclusive (greater than 0).
 The rules are listed at  [protopedia](https://proteopedia.org/wiki/index.php/PDB_code).
+
+PDB IDs can have an underscore and then a single letter or number, our code allows for an underscore at the end and doesn't capture the remaining bits.
+
 
 ---
 
@@ -27,29 +59,23 @@ The rules are listed at  [protopedia](https://proteopedia.org/wiki/index.php/PDB
 
 ---
 
-## Miscellaneous notes
+## Miscellaneous
 
-Possible ID types are listed on the [GenBank File Format](https://www.ncbi.nlm.nih.gov/Sitemap/samplerecord.html).
-
-
----
-
-## Archive
-
-We may want to check in the PMCID full text retrieval system for open access papers. Information on this is [here](https://ftp.ncbi.nlm.nih.gov/pub/pmc/) and [here](https://www.ncbi.nlm.nih.gov/pmc/tools/get-full-text/). One advantage of this approach is that the XML text should be much cleaner and avoid html gobble that happens to be PDBs. Another advantage of this approach is that source files like supplementary data and images are in the open access subset. It seems like there may be a better approach than simply inserting the PMCID in the link like in the code below, such as using a library for direct queries of the database, like [pymed/article.py](https://github.com/gijswobben/pymed/blob/master/pymed/article.py). But this approach is more complicated and requries writing much more code. Also it doesn't seem like the xml text much noise/gobble, so direct queries may not provide much advantage. Some papers like [https://doi.org/10.1515/cclm-2021-1287](10.1515/cclm-2021-1287) are technically on pubmed, it has a PMID, but either aren't on pubmed central and thus don't have a PMCID or aren't open access. This paper doesn't even appear on the `metapub` fetch. When searched by doi, but when searched by its PMID, it's found.
-
+#### PMC full-text retrieval for supplementary information
+We may want to check in the PMCID full text retrieval system for open access papers. Information on this is [here](https://ftp.ncbi.nlm.nih.gov/pub/pmc/) and [here](https://www.ncbi.nlm.nih.gov/pmc/tools/get-full-text/). One advantage of this approach is that source files like supplementary data and images are in the open access subset. One approach is to simply insert the PMCID in a hyperlink, such as in the following code:
 ```
 pmc_oa_xml = "https://eutils.ncbi.nlm.nih.gov/entrez/eutils/efetch.fcgi?db=pmc&id=" + self.pmc
 ```
+An alternative approach may be to use a library for direct queries of the database, like [pymed/article.py](https://github.com/gijswobben/pymed/blob/master/pymed/article.py). But this approach is more complicated as it might requre writing much more code.
 
-We should write a checker to see if this approach fails. The xml page has a standard error messsage when a paper does not exist I think.
+#### Strange papers
+Some papers like [https://doi.org/10.1515/cclm-2021-1287](10.1515/cclm-2021-1287) are technically on pubmed, it has a PMID, but either aren't on pubmed central and thus don't have a PMCID or aren't open access. This paper doesn't even appear on the `metapub` fetch. When searched by doi, but when searched by its PMID, it's found.
 
----
-
-Interesting case study:
-
+#### Interesting case study
 * No doi
 * [PMID method](https://pubmed.ncbi.nlm.nih.gov/34873578/)
 * [PMCID method](https://www.ncbi.nlm.nih.gov/pmc/articles/PMC8647651/)
 * [Weird PMCID requests method](https://eutils.ncbi.nlm.nih.gov/entrez/eutils/efetch.fcgi?db=pmc&id=8647651&tool=my_tool&email=my_email@example.com)
 
+
+---
