+++ data/explore/pubmed_null.csv
@@ -1,0 +1,7 @@
+,0
+title,0
+authors,0
+date,0
+abstract,136
+journal,5
+doi,86
