+++ data/explore/pubmed_null_doi_journals.csv
@@ -1,0 +1,44 @@
+,journal
+ArXiv,10
+"Journal of immunology (Baltimore, Md. : 1950)",9
+Avian diseases,5
+Bing du xue bao = Chinese journal of virology,4
+Zhonghua shi yan he lin chuang bing du xue za zhi = Zhonghua shiyan he linchuang bingduxue zazhi = Chinese journal of experimental and clinical virology,3
+Sheng wu hua xue yu sheng wu wu li xue bao Acta biochimica et biophysica Sinica,3
+American journal of veterinary research,3
+Canadian journal of veterinary research = Revue canadienne de recherche veterinaire,3
+Hong Kong medical journal = Xianggang yi xue za zhi,2
+Sheng wu gong cheng xue bao = Chinese journal of biotechnology,2
+Zhonghua yi xue za zhi,2
+Journal of biosciences,2
+The new microbiologica,2
+Xi bao yu fen zi mian yi xue za zhi = Chinese journal of cellular and molecular immunology,2
+The Japanese journal of veterinary research,1
+Lakartidningen,1
+Biomedical and environmental sciences : BES,1
+DTW. Deutsche tierarztliche Wochenschrift,1
+Wei sheng wu xue bao = Acta microbiologica Sinica,1
+Acta virologica,1
+Pakistan journal of pharmaceutical sciences,1
+BMB reports,1
+Acta naturae,1
+Medicina,1
+Iranian journal of kidney diseases,1
+Journal of biological regulators and homeostatic agents,1
+Beijing da xue xue bao. Yi xue ban = Journal of Peking University. Health sciences,1
+Voprosy virusologii,1
+Archivum immunologiae et therapiae experimentalis,1
+Cellular & molecular immunology,1
+Antiviral therapy,1
+Japanese journal of infectious diseases,1
+Immunology,1
+Journal of the American Veterinary Medical Association,1
+The Journal of biological chemistry,1
+Laboratory animal science,1
+Journal of molecular recognition : JMR,1
+Annals of Ibadan postgraduate medicine,1
+Molecules and cells,1
+Di 1 jun yi da xue xue bao = Academic journal of the first medical college of PLA,1
+Chinese medical journal,1
+Current opinion in molecular therapeutics,1
+Annales de recherches veterinaires. Annals of veterinary research,1
