--- paperscraper/missing_titles
@@ -1,2 +1,0 @@
-['production of an anti-severe acute respiratory syndrome (sars) coronavirus human monoclonal antibody fab fragment by using a combinatorial immunoglobulin gene library derived from patients who recovered from sars.', 'mouse antibodies with activity against the sars-cov-2 d614g and b.1.351 variants.', 'structures of potent and convergent neutralizing antibodies bound to the sars-cov-2 spike unveil a unique epitope responsible for exceptional potency.', 'sars-cov-2 beta variant infection elicits potent lineage-specific and cross-reactive antibodies.', 'fragment-based computational design of antibodies targeting structured epitopes.', 'discovery of nanobodies against sars-cov-2 and an uncommon neutralizing mechanism.', 'broadening a sars-cov-1 neutralizing antibody for potent sars-cov-2 neutralization through directed evolution.', 'single-cell sequencing of plasma cells from covid-19 patients reveals highly expanded clonal lineages produce specific and neutralizing antibodies to sars-cov-2.', 'deep mining of early antibody response in covid-19 patients yields potent neutralisers and reveals high level of convergence.', 'cross-reactivity of antibody against sars-coronavirus nucleocapsid protein with il-11.', 'potent synthetic nanobodies against sars-cov-2 and molecular basis for neutralization.', 'chicken single-chain variable fragments against the sars-cov spike protein.', 'development of sars-cov-2 nucleocapsid specific monoclonal antibodies.', 'human single-chain antibodies neutralize sars-cov-2 variants by engaging an essential epitope of the spike: a new weapon against covid-19.', 'isolation and identification of an scfv antibody against nucleocapsid protein of sars-cov.', 'the development of nanosota-1 as anti-sars-cov-2 nanobody drug candidates.', 'characterization of protease activity of nsp3 from sars-cov-2 and its in vitro inhibition by nanobodies.', 'structure of a germline-like human antibody defines a neutralizing epitope on the sars-cov-2 spike ntd.', 'sybodies targeting the sars-cov-2 receptor-binding domain.', 'efficient discovery of potently neutralizing sars-cov-2 antibodies using libra-seq with ligand blocking.', 'bi-paratopic and multivalent human vh domains neutralize sars-cov-2 by targeting distinct epitopes within the ace2 binding interface of spike.', 'fast isolation of sub-nanomolar affinity alpaca nanobody against the spike rbd of sars-cov-2 by combining bacterial display and a simple single-step density gradient selection.', 'enhanced sars-cov-2 neutralization by dimeric iga.']
-
